Add files via upload
Using 120 dog breeds from Stanford Dogs dataset There are 20,580 images, out of which 12,000 are used for
training and 8580 for testing. Class labels and bounding box annotations are provided for all the 12,000 images.
By turning images into batches and converting to tensors and creating a model that evaluates and predicts the
breed of the dog from the input image. All the predictions are based on the trained datasets